Added Search, Finished comments
Most of the site is now finished this includes:
+ Thread listing homepage
+ View comments on thread (topics) with timestamp
+ Add new comment
+ Add new topics
+ Added descriptions to topics
+ Recent topics (last 3 added topics)
+ Search functionality

Even with all these features should this project be taken further it
would require to have the core functionality rewritten to a HASH based
id system, like an addressable SET. Most likely move it towards a database
and start fresh with a MVC framework like Spring. As a prototype it does
everything i need it to.